♻️ Refactor: dto 수정 및 swagger-ui 적용
