+++ src/main/java/bumblebee/xchangepass/domain/wallet/transaction/dto/WalletTransactionMessage.java
@@ -1,7 +1,8 @@
-package bumblebee.xchangepass.domain.wallet.wallet.dto.request;
+package bumblebee.xchangepass.domain.wallet.transaction.dto;
 
 import java.math.BigDecimal;
 
+//내부 지갑 트랜잭션 메시지
 public record WalletTransactionMessage(
         Long myWalletId,
         Long counterWalletId,
