+++ src/main/java/bumblebee/xchangepass/domain/wallet/wallet/controller/WalletController.java
@@ -1,6 +1,7 @@
 package bumblebee.xchangepass.domain.wallet.wallet.controller;
 
 import bumblebee.xchangepass.domain.wallet.transaction.dto.request.WalletTransactionSearchCondition;
+import bumblebee.xchangepass.domain.wallet.transaction.dto.response.WalletTransactionListResponse;
 import bumblebee.xchangepass.domain.wallet.transaction.service.WalletTransactionService;
 import bumblebee.xchangepass.domain.wallet.wallet.dto.request.WalletInOutRequest;
 import bumblebee.xchangepass.domain.wallet.wallet.dto.request.WalletTransferRequest;
@@ -37,12 +38,18 @@ public class WalletController {
 
     @Operation(summary = "거래내역 조회", description = "거래내역을 조회합니다.")
     @ApiResponses(value = {
+            @ApiResponse(responseCode = "400", description = "거래내역 조회 실패",
+                    content = @Content(mediaType = "application/json",
+                            schema = @Schema(implementation = ErrorResponse.class),
+                            examples = @ExampleObject(value = "{\n  \"code\": \"W001\"," +
+                                                              "\n  \"message\": \"지갑을 찾을 수 없습니다.\"}"))
+            )
     })
     @GetMapping("/transaction")
     @ResponseStatus(HttpStatus.OK)
-    public List<WalletTransactionResponse> transaction(Authentication authentication,
-                                                       @ModelAttribute WalletTransactionSearchCondition condition,
-                                                       Pageable pageable) {
+    public List<WalletTransactionListResponse> transaction(Authentication authentication,
+                                                           @ModelAttribute WalletTransactionSearchCondition condition,
+                                                           Pageable pageable) {
         return transactionService.getTransaction(JwtUtil.getLoginId(authentication), condition, pageable);
     }
